function created to handle dynamic header
