+++ src/components/pages/ParentDashboard/RenderParentDashboard.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { Layout } from 'antd';
 import { useOktaAuth } from '@okta/okta-react';
 import { getProfileData } from '../../../api';
-import ParentNavTopBar from '../../common/ParentNavTop';
+// import ParentNavTopBar from '../../common/ParentNavTop';
 // import NewProgressCharts from '../../common/ProgressChart';
 import NewChildCard from '../../common/ChildCard';
 import AccountSettings from '../ParentAccountSettings/AccountSettingsContainer';
@@ -21,10 +21,11 @@ const RenderNewParentDashboard = props => {
       });
     });
   }, [setParent, authState]);
+
   return (
     <div>
       <Layout className="newparent-dashboard">
-        <ParentNavTopBar />
+        {/* <ParentNavTopBar /> */}
         <Layout>
           <div className="ProgressContainer">
             <div className="ProgressHeader center-content">
